fix(ui): Improve location selector and data builder
Adjusts the display of location names by removing bold styling from the location name in the `LieuSelector` component.

Updates the data builder to ensure a consistent order for DATs at "Oncopole-Lise Enjalbert" and "Université Paul-Sabatier" stations on the TELEO line. This change reverses the order of DATs to be `02 (droite)` then `01 (gauche)`.